requirements.txt
streamlit==1.11.1
matplotlib==3.4.2
seaborn==0.11.1
numpy==1.22.0
catboost
category-encoders
plotly
pillow